refactor: suggestions/corrections
