+++ data/sequences.fasta
@@ -1,0 +1,8 @@
+>sequence1
+GAAGTTTACTAaCTGGAGTGGTCAGaAGTTGCCGCCTGTG
+>sequence2
+GCCCGGGCGTATGTATGAGAGATGTGGCCAGAAGTCGAAA
+>sequence3
+ATGATTDATGTGTCCGGTAACTATAAACGTGCTACGATGT
+>sequence4
+TTTGAG
