+++ setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup, find_packages
+
+setup(
+    name='src',
+    packages=find_packages(),
+    install_requires=[],
+)
